backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["baktiyar <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.43"
pyjwt = "^2.10.1"
alembic = "^1.16.5"
pydantic = {extras = ["email"], version = "^2.11.7"}
passlib = {version = "1.7.4", extras = ["bcrypt"]}
bcrypt = "3.2.2"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
aiosqlite = "^0.21.0"
pytest-cov = "^6.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
